requirements.txt
discord.py>=2.0.0
yt-dlp>=2023.0.0
PyNaCl>=1.5.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
